Added Logger.xml and added log statement in all java files.

+++ EmployeeManagement/src/main/java/in/employeeManagement/service/EmployeeServiceImpl.java
@@ -52,6 +52,10 @@ public class EmployeeServiceImpl implements EmployeeService {
 	@Transactional
 	@Override
 	public void delete(int id) {
+		Employee emp = employeeDAO.getEmployee(id);
+		if(emp == null) {
+			throw new EmployeeNotFoundException("Id does not exist. Enter correct Id!");
+		}
 		employeeDAO.delete(id);
 	}
